tratamento das validações, exceções e testes de execução

+++ src/main/java/com/sumauma/crudOfClients/services/ClientService.java
@@ -1,7 +1,5 @@
 package com.sumauma.crudOfClients.services;
 
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -11,6 +9,9 @@ import org.springframework.transaction.annotation.Transactional;
 import com.sumauma.crudOfClients.DTO.ClientDTO;
 import com.sumauma.crudOfClients.entities.Client;
 import com.sumauma.crudOfClients.repositories.ClientRepository;
+import com.sumauma.crudOfClients.services.customExceptions.NoSuchElementException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 
 @Service
@@ -21,11 +22,9 @@ public class ClientService {
 		
 	@Transactional(readOnly = true)
 	public ClientDTO findById(Long id) {
-		Optional<Client> result = clientRepository.findById(id);
-		Client client = result.get();
-		ClientDTO clientDto = new ClientDTO(client);
-		return clientDto;
-		
+		Client client = clientRepository.findById(id).orElseThrow(
+				() -> new NoSuchElementException("Cliente inexistente"));
+		return new ClientDTO(client);
 	}
 	
 	@Transactional(readOnly = true)
@@ -38,26 +37,30 @@ public class ClientService {
 	public ClientDTO insert(ClientDTO clientDto) {
 		Client client = new Client();
 		copyClientDtoToClient(clientDto, client);
-		client = clientRepository.save(client);
-		return new ClientDTO(client);
-		
+		return new ClientDTO(clientRepository.save(client));
 	}	
 	
 	@Transactional
 	public ClientDTO update(Long id, ClientDTO cleintDto) {
-		Client cleint = clientRepository.getReferenceById(id);
-		copyClientDtoToClient(cleintDto, cleint);
-		cleint = clientRepository.save(cleint);
-		return new ClientDTO(cleint);
+		try {
+			Client cleint = clientRepository.getReferenceById(id);
+			copyClientDtoToClient(cleintDto, cleint);
+			return new ClientDTO(clientRepository.save(cleint));
+		}
+		catch(EntityNotFoundException e) {
+			throw new NoSuchElementException("Cliente inexistente");
+		}
 	}
 	
-	
+	@Transactional
 	public void delete(Long id) {
+		if (!clientRepository.existsById(id)) {
+			throw new NoSuchElementException("Cliente inexistente");
+		}
 		clientRepository.deleteById(id);
 	}
 		
 	private void copyClientDtoToClient(ClientDTO clientDto, Client client) {
-		client.setId(clientDto.getId());
 		client.setName(clientDto.getName());
 		client.setCpf(clientDto.getCpf());
 		client.setIncome(clientDto.getIncome());
